Repot into new dir
