+++ frame/support/src/traits.rs
@@ -38,6 +38,13 @@ use crate::weights::Weight;
 use bitflags::bitflags;
 use impl_trait_for_tuples::impl_for_tuples;
 
+mod fungibles;
+mod fungible;
+pub use fungible::{InspectFungible, Fungible, TransferFungible, ReserveFungible, AssetOf};
+pub use fungibles::{
+	InspectFungibles, Fungibles, TransferFungibles, ReserveFungibles, WithdrawConsequence
+};
+
 /// Re-expected for the macro.
 #[doc(hidden)]
 pub use sp_std::{mem::{swap, take}, cell::RefCell, vec::Vec, boxed::Box};
@@ -1126,235 +1133,6 @@ pub trait Currency<AccountId> {
 		who: &AccountId,
 		balance: Self::Balance,
 	) -> SignedImbalance<Self::Balance, Self::PositiveImbalance>;
-}
-
-/// One of a number of consequences of withdrawing a fungible from an account.
-#[derive(Copy, Clone, Eq, PartialEq)]
-pub enum WithdrawConsequence<Balance> {
-	/// Withdraw could not happen.
-	Failed,
-	/// Account balance would reduce to zero, potentially destroying it. The parameter is the
-	/// amount of balance which is destroyed.
-	ReducedToZero(Balance),
-	/// Account continued in existence.
-	Success,
-}
-
-/// Trait for providing balance-inspection access to a set of named fungible assets.
-pub trait InspectFungibles<AccountId> {
-	/// Means of identifying one asset class from another.
-	type AssetId: FullCodec + Copy + Default;
-	/// Scalar type for representing balance of an account.
-	type Balance: AtLeast32BitUnsigned + FullCodec + Copy + Default;
-	/// The total amount of issuance in the system.
-	fn total_issuance(asset: Self::AssetId) -> Self::Balance;
-	/// The minimum balance any single account may have.
-	fn minimum_balance(asset: Self::AssetId) -> Self::Balance;
-	/// Get the `asset` balance of `who`.
-	fn balance(asset: Self::AssetId, who: &AccountId) -> Self::Balance;
-	/// Returns `true` if the `asset` balance of `who` may be increased by `amount`.
-	fn can_deposit(asset: Self::AssetId, who: &AccountId, amount: Self::Balance) -> bool;
-	/// Returns `Failed` if the `asset` balance of `who` may not be decreased by `amount`, otherwise
-	/// the consequence.
-	fn can_withdraw(
-		asset: Self::AssetId,
-		who: &AccountId,
-		amount: Self::Balance,
-	) -> WithdrawConsequence<Self::Balance>;
-}
-
-/// Trait for providing a set of named fungible assets which can be created and destroyed.
-pub trait Fungibles<AccountId>: InspectFungibles<AccountId> {
-	/// Increase the `asset` balance of `who` by `amount`.
-	fn deposit(asset: Self::AssetId, who: &AccountId, amount: Self::Balance) -> DispatchResult;
-	/// Attempt to reduce the `asset` balance of `who` by `amount`.
-	fn withdraw(asset: Self::AssetId, who: &AccountId, amount: Self::Balance) -> DispatchResult;
-	/// Transfer funds from one account into another.
-	fn transfer(
-		asset: Self::AssetId,
-		source: &AccountId,
-		dest: &AccountId,
-		amount: Self::Balance,
-	) -> DispatchResult {
-		if !Self::can_deposit(asset, &dest, amount) {
-			return Err(DispatchError::Other("Cannot deposit"))
-		}
-		Self::withdraw(asset, source, amount)?;
-		let result = Self::deposit(asset, dest, amount);
-		debug_assert!(result.is_ok(), "can_deposit returned true for a failing deposit!");
-		result
-	}
-}
-
-/// Trait for providing a set of named fungible assets which can only be transferred.
-pub trait TransferableFungibles<AccountId>: InspectFungibles<AccountId> {
-	/// Transfer funds from one account into another.
-	fn transfer(
-		asset: Self::AssetId,
-		source: &AccountId,
-		dest: &AccountId,
-		amount: Self::Balance,
-	) -> DispatchResult;
-}
-
-/// Trait for providing a set of named fungible assets which can be reserved.
-pub trait ReservableFungibles<AccountId>: InspectFungibles<AccountId> {
-	/// Amount of funds held in reserve.
-	fn reserved_balance(asset: Self::AssetId, who: &AccountId) -> Self::Balance;
-
-	/// Amount of funds held in reserve.
-	fn total_balance(asset: Self::AssetId, who: &AccountId) -> Self::Balance;
-
-	/// Check to see if some `amount` of `asset` may be reserved on the account of `who`.
-	fn can_reserve(asset: Self::AssetId, who: &AccountId, amount: Self::Balance) -> bool;
-
-	/// Reserve some funds in an account.
-	fn reserve(asset: Self::AssetId, who: &AccountId, amount: Self::Balance) -> DispatchResult;
-
-	/// Unreserve some funds in an account.
-	fn unreserve(asset: Self::AssetId, who: &AccountId, amount: Self::Balance) -> DispatchResult;
-
-	/// Transfer reserved funds into another account.
-	fn repatriate_reserved(
-		asset: Self::AssetId,
-		who: &AccountId,
-		amount: Self::Balance,
-		status: BalanceStatus,
-	) -> DispatchResult;
-}
-
-/// Trait for providing balance-inspection access to a fungible asset.
-pub trait InspectFungible<AccountId> {
-	/// Scalar type for representing balance of an account.
-	type Balance: AtLeast32BitUnsigned + FullCodec + Copy + Default;
-	/// The total amount of issuance in the system.
-	fn total_issuance() -> Self::Balance;
-	/// The minimum balance any single account may have.
-	fn minimum_balance() -> Self::Balance;
-	/// Get the balance of `who`.
-	fn balance(who: &AccountId) -> Self::Balance;
-	/// Returns `true` if the balance of `who` may be increased by `amount`.
-	fn can_deposit(who: &AccountId, amount: Self::Balance) -> bool;
-	/// Returns `Failed` if the balance of `who` may not be decreased by `amount`, otherwise
-	/// the consequence.
-	fn can_withdraw(who: &AccountId, amount: Self::Balance) -> WithdrawConsequence<Self::Balance>;
-}
-/// Trait for providing an ERC-20 style fungible asset.
-pub trait Fungible<AccountId>: InspectFungible<AccountId> {
-	/// Increase the balance of `who` by `amount`.
-	fn deposit(who: &AccountId, amount: Self::Balance) -> DispatchResult;
-	/// Attempt to reduce the balance of `who` by `amount`.
-	fn withdraw(who: &AccountId, amount: Self::Balance) -> DispatchResult;
-}
-/// Trait for providing a fungible asset which can only be transferred.
-pub trait TransferableFungible<AccountId>: InspectFungible<AccountId> {
-	/// Transfer funds from one account into another.
-	fn transfer(
-		source: &AccountId,
-		dest: &AccountId,
-		amount: Self::Balance,
-	) -> DispatchResult;
-}
-/// Trait for providing a fungible asset which can be reserved.
-pub trait ReservableFungible<AccountId>: InspectFungible<AccountId> {
-	/// Amount of funds held in reserve by `who`.
-	fn reserved_balance(who: &AccountId) -> Self::Balance;
-	/// Amount of funds held in total by `who`.
-	fn total_balance(who: &AccountId) -> Self::Balance {
-		Self::reserved_balance(who).saturating_add(Self::balance(who))
-	}
-	/// Check to see if some `amount` of funds may be reserved on the account of `who`.
-	fn can_reserve(who: &AccountId, amount: Self::Balance) -> bool;
-	/// Reserve some funds in an account.
-	fn reserve(who: &AccountId, amount: Self::Balance) -> DispatchResult;
-	/// Unreserve some funds in an account.
-	fn unreserve(who: &AccountId, amount: Self::Balance) -> DispatchResult;
-	/// Transfer reserved funds into another account.
-	fn repatriate_reserved(
-		who: &AccountId,
-		amount: Self::Balance,
-		status: BalanceStatus,
-	) -> DispatchResult;
-}
-
-pub struct AssetOf<
-	F: InspectFungibles<AccountId>,
-	A: Get<<F as InspectFungibles<AccountId>>::AssetId>,
-	AccountId,
->(
-	sp_std::marker::PhantomData<(F, A, AccountId)>
-);
-impl<
-	F: InspectFungibles<AccountId>,
-	A: Get<<F as InspectFungibles<AccountId>>::AssetId>,
-	AccountId,
-> InspectFungible<AccountId> for AssetOf<F, A, AccountId> {
-	type Balance = <F as InspectFungibles<AccountId>>::Balance;
-	fn total_issuance() -> Self::Balance {
-		<F as InspectFungibles<AccountId>>::total_issuance(A::get())
-	}
-	fn minimum_balance() -> Self::Balance {
-		<F as InspectFungibles<AccountId>>::minimum_balance(A::get())
-	}
-	fn balance(who: &AccountId) -> Self::Balance {
-		<F as InspectFungibles<AccountId>>::balance(A::get(), who)
-	}
-	fn can_deposit(who: &AccountId, amount: Self::Balance) -> bool {
-		<F as InspectFungibles<AccountId>>::can_deposit(A::get(), who, amount)
-	}
-	fn can_withdraw(who: &AccountId, amount: Self::Balance) -> WithdrawConsequence<Self::Balance> {
-		<F as InspectFungibles<AccountId>>::can_withdraw(A::get(), who, amount)
-	}
-}
-
-impl<
-	F: Fungibles<AccountId>,
-	A: Get<<F as InspectFungibles<AccountId>>::AssetId>,
-	AccountId,
-> Fungible<AccountId> for AssetOf<F, A, AccountId> {
-	fn deposit(who: &AccountId, amount: Self::Balance) -> DispatchResult {
-		<F as Fungibles<AccountId>>::deposit(A::get(), who, amount)
-	}
-	fn withdraw(who: &AccountId, amount: Self::Balance) -> DispatchResult {
-		<F as Fungibles<AccountId>>::withdraw(A::get(), who, amount)
-	}
-}
-impl<
-	F: TransferableFungibles<AccountId>,
-	A: Get<<F as InspectFungibles<AccountId>>::AssetId>,
-	AccountId,
-> TransferableFungible<AccountId> for AssetOf<F, A, AccountId> {
-	fn transfer(source: &AccountId, dest: &AccountId, amount: Self::Balance) -> DispatchResult {
-		<F as TransferableFungibles<AccountId>>::transfer(A::get(), source, dest, amount)
-	}
-}
-impl<
-	F: ReservableFungibles<AccountId>,
-	A: Get<<F as InspectFungibles<AccountId>>::AssetId>,
-	AccountId,
-> ReservableFungible<AccountId> for AssetOf<F, A, AccountId> {
-	fn reserved_balance(who: &AccountId) -> Self::Balance {
-		<F as ReservableFungibles<AccountId>>::reserved_balance(A::get(), who)
-	}
-	fn total_balance(who: &AccountId) -> Self::Balance {
-		<F as ReservableFungibles<AccountId>>::total_balance(A::get(), who)
-	}
-	fn can_reserve(who: &AccountId, amount: Self::Balance) -> bool {
-		<F as ReservableFungibles<AccountId>>::can_reserve(A::get(), who, amount)
-	}
-	fn reserve(who: &AccountId, amount: Self::Balance) -> DispatchResult {
-		<F as ReservableFungibles<AccountId>>::reserve(A::get(), who, amount)
-	}
-	fn unreserve(who: &AccountId, amount: Self::Balance) -> DispatchResult {
-		<F as ReservableFungibles<AccountId>>::unreserve(A::get(), who, amount)
-	}
-	fn repatriate_reserved(
-		who: &AccountId,
-		amount: Self::Balance,
-		status: BalanceStatus,
-	) -> DispatchResult {
-		<F as ReservableFungibles<AccountId>>::repatriate_reserved(A::get(), who, amount, status)
-	}
 }
 
 /// Status of funds.
